Format Makefile (1/2): reorder lines
Format the Makefile to be more readible and easier to navigate for the
first-time reader. This way, the Makefile becomes its own documentation.
The general layout is:
 * General configuration
 * Internal variables
 * Global rules
 * Free/Libre Tools
 * Non-free SDK

This step brings all statements (variable assignments and rule-blocks) in
the Makefile in a new order. No further edits were done. One possible
sanity check is:

diff <(sort Makefile.orig) <(sort Makefile.new)