Makefile: split up .PHONY rule esptool
Split up the .PHONY rule with side effect (the copy command) into two
rules: one "normal" rule for the destination file of the copy command and
one .PHONY rule without side effect.

This fixes the problem that "make all" executed the copy even if the
destination file already existed and was current.