Makefile, patch 1.2.0: fix commented out patches
Comment out the unused dependency for the SDK patch of version 1.2.0.

Also add the (also unused but currently unfulfillable) dependency of
libsmartconfig_2.4.2.zip as commented out: this is a reminder in case
the file ever gets used again (with a mirror URL).

And order the commands in that rule to group the still executed patches
and commented out ones better.