Add new header esp_protos.h to all SDKs
Several functions used by the public interface of the SDK are not
properly declared in the headers. So these are implicitly declared on
first use and no checking on the call parameters (number and type) or
return value type is done.

Other projects have a espmissingincludes.h header that (among others)
lists these prototypes. But there are many different versions in
different projects and there seems to be no central source where to
send patches etc. So adding a similar file here as the central source
for it seems appropriate.

The file is not based of off the various espmissingincludes.h, though
it was compared to those. The file is based on the output of an
automated script that parses the ld/eagle.rom.addr.v6.ld and all
lib/*.a files to get a list of functions and then compares against the
headers in include to find used functions without a propper declaration.
The prototypes given are either taken from the SDK API Guide, adapted
from the standard libc equivalent or guessed...