Introduce common makefile to simplify firmware builds.
Inspired by the common.mk and parameters.mk build mechanism of the
esp-open-rtos project, a similar build mechanism for the esp-open-sdk
project has been created. This Makefile looks up sources files
recursively, builds and links them. Afterwards, it removes unreferenced
code again in order to keep the firmware as small as possible.
Furthermore, all parameters from parameters.mk can be overwritten from
the firmware project's Makefile.

This is possibly not the perfect way to do that but it can be understood
as a good (and for C working) point so start with. For C++, there is
further work to do, the current version will not work yet. Also, it
could be discussed if other attributes like the CFLAGS or LDFLAGS
parameters should be overwritable or extendable.

There will be two more pull requests based on this branch. The first one
will be the modified Makefile for the blinky example, the second one
will be a new example for serial printing with UART. These pull requests
should be used to understand, how the common.mk Makefile can be used.