Makefile: add infos and patch for esptool.py version conflicts
This patch checks which version of python is installed as standard version and if a python2 version exist. It informs the user that esptool.py needs to run under python 2 and gives, related to the above findings advices what to do next. Finally it offers to patch the file for the user by providing an addtional make target.