Fix to get expat-2.1.0.tar.gz from a different location, since it is no longer available from the one that had been used.
NOTE: The reason the old location didn't work is that expat-2.1.0.tar.gz on sourceforge had been renamed to "expat-2.1.0-RENAMED-VULNERABLE-PLEASE-USE-2.3.0-INSTEAD.tar.gz"... I considered seeing if it would work using 2.3.0 as suggested, but 2.3.0 similarly redirects to 2.4.1, and the immediate goal is to get this stuff to build with minimally intrusive changes.

However, it would probably be wise to switch to a newer expat!